requirements.txt
# Dependencias

# Base de datos
mysql-connector-python==8.2.0

# Procesamiento de imágenes
Pillow==10.1.0

# Utilidades adicionales (opcionales)
python-dateutil==2.8.2

# Para desarrollo y testing (opcional)
pytest==7.4.3
pytest-mock==3.12.0

# Documentación
sphinx==7.2.6